refactor: extract pipeline/discovery.rs
Move SSOT layout introspection helpers (is_item_dir, is_bundle_file,
find_bundle_by_name, has_matching_items, scan_source_items,
find_registry_root, has_ssot_layout, detect_item_entrypoint,
iter_item_dirs) into a focused submodule. These functions classify
filesystem paths without touching the parser — they only check
directory structure.

Their inline unit tests move into discovery.rs as well, so each test
lives next to the function under test.

Refs #180
